change install path to @loader_path
This should theoretically still work in a "main App Bundle's Frameworks
forlder" AS WELL as enable it to be a "-sub_umbrella" in another
framework..  The "bundler", aka the app or the main framework then just
specifies it's "Installation Directory" as @rpath…  sets "runpath"  to
"@loader_path/Frameworks", and adds a copy phase for the framework.  It
also may be necessary to add the "-sub_umbrella TwUI" flag to
"reexport" the TwUI classes to "siblings".